requirements.txt
tensorflow
gdown
numpy
scikit-learn
pandas
seaborn
matplotlib
Pillow
fastapi
uvicorn
nest_asyncio
pyngrok
requests
gradio